+++ Rendering/Audio/PublicUtility/CAVolumeCurve.h
@@ -78,12 +78,12 @@ inline bool	operator>(const CARawPoint& x, const CARawPoint& y) { return !((x < 
 
 struct CADBPoint
 {
-	Float64	mMinimum;
-	Float64	mMaximum;
+	Float32	mMinimum;
+	Float32	mMaximum;
 	
 	CADBPoint() : mMinimum(0), mMaximum(0) {}
 	CADBPoint(const CADBPoint& inPoint) : mMinimum(inPoint.mMinimum), mMaximum(inPoint.mMaximum) {}
-	CADBPoint(Float64 inMinimum, Float64 inMaximum) : mMinimum(inMinimum), mMaximum(inMaximum) {}
+	CADBPoint(Float32 inMinimum, Float32 inMaximum) : mMinimum(inMinimum), mMaximum(inMaximum) {}
 	CADBPoint&	operator=(const CADBPoint& inPoint) { mMinimum = inPoint.mMinimum; mMaximum = inPoint.mMaximum; return *this; }
 	
 	static bool	Overlap(const CADBPoint& x, const CADBPoint& y) { return (x.mMinimum < y.mMaximum) && (x.mMaximum >= y.mMinimum); }
@@ -136,8 +136,8 @@ public:
 	void			SetTag(UInt32 inTag)	{ mTag = inTag; }
 	SInt32			GetMinimumRaw() const;
 	SInt32			GetMaximumRaw() const;
-	Float64			GetMinimumDB() const;
-	Float64			GetMaximumDB() const;
+	Float32			GetMinimumDB() const;
+	Float32			GetMaximumDB() const;
 	
 	void			SetIsApplyingTransferFunction(bool inIsApplyingTransferFunction)  { mIsApplyingTransferFunction = inIsApplyingTransferFunction; }
 	int				GetTransferFunction() const { return mTransferFunction; }
@@ -145,16 +145,16 @@ public:
 
 //	Operations
 public:
-	void			AddRange(SInt32 mMinRaw, SInt32 mMaxRaw, Float64 inMinDB, Float64 inMaxDB);
+	void			AddRange(SInt32 mMinRaw, SInt32 mMaxRaw, Float32 inMinDB, Float32 inMaxDB);
 	void			ResetRange();
 	bool			CheckForContinuity() const;
 	
-	SInt32			ConvertDBToRaw(Float64 inDB) const;
-	Float64			ConvertRawToDB(SInt32 inRaw) const;
-	Float64			ConvertRawToScalar(SInt32 inRaw) const;
-	Float64			ConvertDBToScalar(Float64 inDB) const;
-	SInt32			ConvertScalarToRaw(Float64 inScalar) const;
-	Float64			ConvertScalarToDB(Float64 inScalar) const;
+	SInt32			ConvertDBToRaw(Float32 inDB) const;
+	Float32			ConvertRawToDB(SInt32 inRaw) const;
+	Float32			ConvertRawToScalar(SInt32 inRaw) const;
+	Float32			ConvertDBToScalar(Float32 inDB) const;
+	SInt32			ConvertScalarToRaw(Float32 inScalar) const;
+	Float32			ConvertScalarToDB(Float32 inScalar) const;
 
 //	Implementation
 private:
@@ -164,8 +164,8 @@ private:
 	CurveMap		mCurveMap;
 	bool			mIsApplyingTransferFunction;
 	UInt32			mTransferFunction;
-	Float64			mRawToScalarExponentNumerator;
-	Float64			mRawToScalarExponentDenominator;
+	Float32			mRawToScalarExponentNumerator;
+	Float32			mRawToScalarExponentDenominator;
 
 };
