+++ Utilities/VirtualRingBuffer.h
@@ -57,7 +57,7 @@
 {
     void* buffer;
     void* bufferEnd;
-    UInt32 bufferLength;
+    size_t bufferLength;
         // buffer is the start of the ring buffer's address space.
         // bufferEnd is the end of the "real" buffer (always buffer + bufferLength).
         // Note that the "virtual" portion of the buffer extends from bufferEnd to bufferEnd+bufferLength.
@@ -67,7 +67,7 @@
 }
 
 // Note: The specified length will be rounded up to an integral number of VM pages.
-- (id)initWithLength:(UInt32)length;
+- (id)initWithLength:(size_t)length;
 
 // Empties the buffer. It is NOT safe to do this while anyone is reading from or writing to the buffer.
 - (void)empty;
@@ -77,17 +77,17 @@
 // Read operations:
 
 // The reading thread must call this method first.
-- (UInt32)lengthAvailableToReadReturningPointer:(void **)returnedReadPointer;
+- (size_t)lengthAvailableToReadReturningPointer:(void **)returnedReadPointer;
 // Iff a value > 0 is returned, the reading thread may go on to read that much data from the returned pointer.
 // Afterwards, the reading thread must call didReadLength:.
-- (void)didReadLength:(UInt32)length;
+- (void)didReadLength:(size_t)length;
 
 // Write operations:
 
 // The writing thread must call this method first.
-- (UInt32)lengthAvailableToWriteReturningPointer:(void **)returnedWritePointer;
+- (size_t)lengthAvailableToWriteReturningPointer:(void **)returnedWritePointer;
 // Iff a value > 0 is returned, the writing thread may then write that much data into the returned pointer.
 // Afterwards, the writing thread must call didWriteLength:.
-- (void)didWriteLength:(UInt32)length;
+- (void)didWriteLength:(size_t)length;
 
 @end
